[moe] refactor moe routers
